chore: refactor tce proxy, subnet runtime and sequencer types
chore: refactor topos-sequencer-tce-proxy
chore: refactor topos-sequencer-subnet-runtime-proxy
chore: remove topos-sequencer-types
chore: rename modules
