+++ .github/workflows/sequencer_topos_core_contract_test.yml
@@ -74,5 +74,5 @@ jobs:
       - name: Run all workspace tests
         run: |
           export TOPOS_SMART_CONTRACTS_BUILD_PATH=`pwd`/contracts
-          cd crates/topos-sequencer-subnet-runtime-proxy
+          cd crates/topos-sequencer-subnet-runtime
           cargo test -- --nocapture
