--- crates/topos-sequencer-types/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "topos-sequencer-types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-topos-core = { workspace = true, features = ["uci", "api"] }
-tokio.workspace = true
-thiserror.workspace = true
-ethereum-types.workspace = true
-serde = { workspace = true, features = ["derive"] }
-tracing.workspace = true
-opentelemetry.workspace = true
